Add equidistant distortion model
This is a forward-port from ROS1 noetic:

https://github.com/ros/common_msgs/pull/109

The model is based on the following publication:

J. Kannala and S. Brandt (2006). A Generic Camera Model and Calibration
Method for Conventional, Wide-Angle, and Fish-Eye Lenses, IEEE
Transactions on Pattern Analysis and Machine Intelligence, vol. 28, no.
8, pp. 1335-1340

There are many different names for this distortion model:

* Kannala and Brandt call it "Generic Camera Model". Sometimes they also
  call it "equidistance projection model".
* Kalibr calls it "equidistant": https://github.com/ethz-asl/kalibr/wiki/supported-models
* Camodocal calls it "kannala-brandt": https://github.com/hengli/camodocal
* OpenCV calls it "fisheye": https://stackoverflow.com/a/34309644/3036576

All of these are just different names for the same model.

Signed-off-by: Martin Günther <martin.guenther@dfki.de>