Update common_interfaces to C++17. (#215)
The main reason to do this is so that we can compile
common_interfaces with the clang static analyzer. As of
clang++-14 (what is in Ubuntu 22.04), the default still
seems to be C++14, so we need to specify C++17 so that
new things in the rclcpp headers work properly.

Further, due to reasons I don't fully understand, I needed
to set CMAKE_CXX_STANDARD_REQUIRED in order for clang to
really use that version. So set this as well.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>